refactor: move shared repository logic to provider/shared
