+++ internal/core/services/public_cloud/service_test.go
@@ -11,8 +11,8 @@ import (
 	"github.com/leaseweb/leaseweb-go-sdk/publicCloud"
 	shared2 "github.com/leaseweb/terraform-provider-leaseweb/internal/core/shared"
 	"github.com/leaseweb/terraform-provider-leaseweb/internal/provider/publiccloud/contracts"
-	"github.com/leaseweb/terraform-provider-leaseweb/internal/provider/publiccloud/shared/repository"
 	"github.com/leaseweb/terraform-provider-leaseweb/internal/provider/resources/public_cloud/model"
+	"github.com/leaseweb/terraform-provider-leaseweb/internal/provider/shared/repository"
 	"github.com/stretchr/testify/assert"
 )
