+++ internal/provider/shared/logging/service_error_test.go
@@ -5,7 +5,7 @@ import (
 	"testing"
 
 	"github.com/hashicorp/terraform-plugin-framework/diag"
-	"github.com/leaseweb/terraform-provider-leaseweb/internal/provider/publiccloud/shared/repository"
+	"github.com/leaseweb/terraform-provider-leaseweb/internal/provider/shared/repository"
 	"github.com/stretchr/testify/assert"
 )
