+++ internal/provider/shared/repository/error.go
