+++ internal/provider/shared/repository/error_test.go
